pyproject.toml
[tool.poetry]
name = "point-cloud-prediction"
version = "0.1.0"
description = ""
authors = ["Benedikt Mersch <mersch@igg.uni-bonn.de>"]
packages = [
{include = "./pcf"},
{include = "./pyTorchChamferDistance"},
]

[tool.poetry.dependencies]
python = "^3.8"
PyYAML = "^6.0"
matplotlib = "^3.4.3"
open3d = "^0.13.0"
pytorch-lightning = "^1.5.0"
ninja = "^1.10.2"


[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
